move function to delayed.js
