+++ scripts/delayed.js
@@ -1,5 +1,7 @@
 // eslint-disable-next-line import/no-cycle
 import { sampleRUM } from './lib-franklin.js';
+// eslint-disable-next-line import/no-cycle
+import { getCookie, setCookie } from './scripts.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
@@ -28,6 +30,34 @@ function LoadDriftWidget() {
     drift.load('pyupvvckemp9');
 };
 /* eslint-enable */
+
+// IPStack Integration to get specific user information
+async function loadUserData() {
+  const countryCodeInfo = 'country_code';
+  if (getCookie(countryCodeInfo)) return;
+  fetch('http://api.ipstack.com/check?access_key=fa0c43f899d86d91bf5aa529a5774566', {
+    /* Referrer Policy is set to strict-origin-when-cross-origin by default
+       to avoid data leaking of private information.
+       See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy#directives
+       Set explicitely policy 'no-referrer-when-downgrade'.
+    */
+    referrerPolicy: 'no-referrer-when-downgrade',
+  }).then((response) => {
+    if (response.ok) {
+      return response.json();
+    }
+    return Promise.reject(response);
+  }).then((data) => {
+    if (data[countryCodeInfo]) {
+      setCookie(countryCodeInfo, data[countryCodeInfo], 30);
+    }
+  }).catch((err) => {
+    // eslint-disable-next-line no-console
+    console.warn('Could not load user information.', err);
+  });
+}
+
+loadUserData();
 
 if (!document.location.hostname.match('www.moleculardevices.com.cn') && !document.location.hostname.match('.hlx.page')) {
   LoadDriftWidget();
